+++ src/package/themes/default/index.ts
@@ -3,6 +3,7 @@ import button from './components/button';
 import buttonGroup from './components/button-group';
 import card from './components/card';
 import container from './components/container';
+import section from './components/section';
 import input from './components/input';
 import inputGroup from './components/input-group';
 import inlineEdit from './components/inline-edit';
@@ -82,6 +83,7 @@ export default {
         buttonGroup,
         card,
         container,
+        section,
         input,
         inputGroup,
         inlineEdit,